pyproject.toml
[project]
name = "adobe-genai-platform"
version = "2.0.0"
description = "AI-powered creative automation platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic[email]>=2.5.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.1",
    "Pillow>=10.3.0",
    "psutil>=5.9.0",
    "PyMuPDF>=1.23.8",
    "python-docx>=1.1.0",
    "PyYAML>=6.0.1",
    "click>=8.1.7",
    "structlog>=24.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "arq>=0.26.0",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.7",
    "boto3>=1.34.0",
    "aioboto3>=13.0.0",
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--asyncio-mode=auto",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
    "performance: marks tests as performance tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit (security)
    "T20", # flake8-print (no print statements)
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "S101",  # assert used (needed in tests)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "T20"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
